cloudinary added , some changes , todo : testing ...

+++ node-server/routes/controllers/categories-controller.js
@@ -1,6 +1,7 @@
 import categoriesModel from "../../models/categories-model.js";
 import ErrorHandler from "../../utils/error-handler.js";
 import asyncErrorHandler from "../../middlewares/async-error-handler.js";
+import stationsModels from "../../models/stations-models.js";
 
 export const addCategory = asyncErrorHandler(async (req, res, next) => {
 
@@ -41,6 +42,13 @@ export const updateCategoryById = asyncErrorHandler(async (req, res, next) => {
 export const deleteCategoryById = asyncErrorHandler(async (req, res, next) => {
 
     const id = req.params.id;
+
+    const station = await stationsModels.find({ category: id });
+
+    if (station.length > 0) {
+        return next(new ErrorHandler("This category can not be deleted as it has stations", 400));
+    }
+
 
     const category = await categoriesModel.findByIdAndDelete(id);
